% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_ggplot.R
\name{calendar_ggplot}
\alias{calendar_ggplot}
\title{ggplot2 version of calendar plot}
\usage{
calendar_ggplot(data, fill, wd, ws, draw.arrow = TRUE)
}
\arguments{
\item{data}{Data frame containing a \code{date} column}

\item{fill}{The column to fill the calendar plot.}

\item{wd, ws}{Optional wind speed and wind direction plotted as arrows.}

\item{draw.arrow}{Should the wind vector arrows be drawn. This is better turned off
if the plot is going to be converted to plotly.}
}
\description{
A ggplot2 equivalent of \code{\link[openair]{calendarPlot}}
}
\details{
Note that the ggplot2 object returned will not have any scale
defined for the fill, so that can be added to the returned
ggplot2 object.
}
\note{
If the dates span multiple years, the year will be appended
to the month name automatically, otherwise, it will not appear.
}
\examples{
{


}
}
