% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{console_log}
\alias{console_log}
\title{Console log based on environment}
\usage{
console_log(x, log_type = "none", .envir = parent.frame(), sys_call = -1)
}
\arguments{
\item{x}{A character vector.

For block styles, conditions, and questions, each element of the
vector becomes a line, and the result is processed by \code{\link[glue:glue]{glue::glue()}}.
For inline styles, each element of the vector becomes an entry in a
comma separated list.}

\item{log_type}{A character. This parameter will look for these options \code{c("info", "warn", "done", "error")} and will prefix the message, \code{x}.}

\item{.envir}{Used to ensure that \code{\link[glue:glue]{glue::glue()}} gets the correct
environment. For expert use only.}

\item{sys_call}{A numeric. Used to ensure \code{sys.call} looks at the right environment. For expert use only.}

\item{display_function_name}{A logical. If \code{TRUE}, this will print out the calling function's name in the logging.}
}
\description{
Console log based on environment
}
