% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_preprocessing.R
\name{coxph_preprocessing}
\alias{coxph_preprocessing}
\title{preprocessing for coxph function}
\usage{
coxph_preprocessing(data, time, status)
}
\arguments{
\item{data, }{data matrix used for analysis, each row is a data vector, or categorical values}

\item{time, }{survival time with same length with data}

\item{status, }{survival status with same length with data}
}
\value{
results
}
\description{
preprocessing for coxph function
}
