% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.coxph.R
\name{drcoxph}
\alias{drcoxph}
\title{coxph function}
\usage{
drcoxph(job_dir, data, time, status, kmtop = 0.5, outformat = "json",
  min_samples = 5, min_groups = 2, pvalcorrection = "fdr")
}
\arguments{
\item{data, }{data matrix used for analysis, each row is a data vector, or categorical values}

\item{time, }{survival time with same length with data}

\item{status, }{survival status with same length with data}

\item{min_samples, }{minimum number of samples for KM analysis}

\item{pvalcorrection, }{fdr, bonferroni}
}
\value{
results
}
\description{
coxph function
}
