% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.diff.R
\name{drdiff}
\alias{drdiff}
\title{Differential function
this function will calculate differentail of all data based on group}
\usage{
drdiff(datafiles, job_dir, outformat = "json", method = "rank-test",
  alternative = "two.sided", min_samples = 5, pvalcorrection = "fdr")
}
\arguments{
\item{datafiles}{contain file1, data matrix used for analysis, each column is a data vector
and file2, group information for each patient}

\item{method, }{t-test, rank-test, (wilcox-test, kruskal, default test)}

\item{alternative, }{two.sided, less, greater (default two.sided)}

\item{min_samples, }{minimum samples for correlation calculation}

\item{pvalcorrection, }{fdr, bonferroni}
}
\value{
results
}
\description{
Differential function
this function will calculate differentail of all data based on group
}
