% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.kaplanmeier.R
\name{drkaplanmeier}
\alias{drkaplanmeier}
\title{Kaplan-Merier function}
\usage{
drkaplanmeier(time, status, group, min_samples = 5, pvalcorrection = "fdr")
}
\arguments{
\item{time, }{survival time with same length with data}

\item{status, }{survival status with same length with data, 1 = Dead, 0 = Alive}

\item{group, }{vector indicating the group information}

\item{min_samples, }{minimum number of samples for KM analysis}

\item{pvalcorrection, }{fdr, bonferroni}
}
\value{
results
}
\description{
Kaplan-Merier function
}
