% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.mutWaterfall.R
\name{drmutWaterfall}
\alias{drmutWaterfall}
\title{mutWaterfall function}
\usage{
drmutWaterfall(mafile, job_dir, format = "json")
}
\arguments{
\item{mafile, }{maf file name}
}
\value{
results
}
\description{
mutWaterfall function
}
\examples{
drmutWaterfall(mafile, "/Users/yumengw/projects/drright/drright-modules/")
}
