% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.oncoplot.R
\name{droncoplot}
\alias{droncoplot}
\title{oncoplot function}
\usage{
droncoplot(mafile, job_dir, format = "json")
}
\arguments{
\item{mafile, }{maf file name}
}
\value{
results
}
\description{
oncoplot function
}
\examples{
drmutWaterfall(mafile, "/Users/yumengw/projects/drright/drright-modules/")
}
