% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_utils.R
\name{drop_nulls}
\alias{drop_nulls}
\title{Removes the null from a vector}
\usage{
drop_nulls(x)
}
\arguments{
\item{x}{a vector to drop the nulls from}
}
\description{
Removes the null from a vector
}
\examples{
drop_nulls(list(1, NULL, 2))
}
