% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{factor_df}
\alias{factor_df}
\title{Convert all characters to factors in a data frame}
\usage{
factor_df(df)
}
\arguments{
\item{df}{Data frame.}
}
\description{
Convert all characters to factors in a data frame
}
