% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_push_all}
\alias{git_push_all}
\title{Push Up Time}
\usage{
git_push_all(message = "update all via git_push_all()")
}
\arguments{
\item{message}{A character. Message to attach to all files being pushed.}

\item{tag}{A logical. If TRUE, this will ask to tag the push with the current package version.}
}
\description{
Only run this function if you are 100\% aware of what you are pushing up! Make sure no client sensitive information is not staged. To avoid pushing up certain files update the \code{.gitignore} file.
}
