% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_utils.R
\name{\%||\%}
\alias{\%||\%}
\title{If x is \code{NULL}, return y, otherwise return x}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Two elements to test, one potentially \code{NULL}}
}
\description{
If x is \code{NULL}, return y, otherwise return x
}
\examples{
NULL \%||\% 1
}
