% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ld_rda}
\alias{ld_rda}
\title{A simplified load function using load().}
\usage{
ld_rda(x, path = "../data", force = F)
}
\arguments{
\item{x}{The object to be loaded.}

\item{force}{Whether the object should be reloaded if it's already in the current environment.}

\item{folder}{To which subfolder the data is saved. If '.', save to current directory.}
}
\description{
A simplified load function using load().
}
