% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{log_info}
\alias{log_info}
\title{Log Info}
\usage{
log_info(x, .envir = parent.frame())
}
\arguments{
\item{x}{A character vector.

For block styles, conditions, and questions, each element of the
vector becomes a line, and the result is processed by \code{\link[glue:glue]{glue::glue()}}.
For inline styles, each element of the vector becomes an entry in a
comma separated list.}

\item{.envir}{Used to ensure that \code{\link[glue:glue]{glue::glue()}} gets the correct
environment. For expert use only.}
}
\description{
Wrapper function for \code{console_log}.
}
