% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_for_py.R
\name{names_for_py}
\alias{names_for_py}
\title{A simplified save function using saveRDS().}
\usage{
names_for_py(dt)
}
\arguments{
\item{dt}{The data.table whose variable names to be changed}
}
\description{
A simplified save function using saveRDS().
}
