% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lmodel2.R
\name{plot_lmodel2}
\alias{plot_lmodel2}
\title{Scatter plot with ggplot2}
\usage{
plot_lmodel2(data, x, y, method = "OLS", mod.line = F, group = NA,
  same.panel = TRUE, xpos = 0.3, ypos = 0.8, pt.col = "blue",
  pt.shape = 1, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x, y}{What to plot as x and y variable.}

\item{method}{Which regression analysis to calculate. There are three methods 
to choose: "OLS"(ordinary least square), "MA"(Major Axis) and "SMA"
(Standardised Major Axis or known as Reduced Major Axis). There are 
calculated using the \code{\link[lmodel2]{lmodel2}} function. Default is "OLS".}

\item{mod.line}{Should 1:1, 2:1 and 1:2 lines be drawn. Default is \code{FALSE}.}

\item{group}{The column name by which the data should be grouped.}

\item{same.panel}{Should all the groups be plotted in the same panel. 
Default is \code{TRUE}. If not they will be separated in different facets as 
in \code{ggplot2}.}

\item{xpos, ypos}{The relative x and y positions where the regression equations 
should be added. This should be a value between 0 and 1. This value multiples
the maximum of the x or y limit gives the (x, y) coordinates of the annotation.}

\item{pt.col, pt.shape}{The colour and shape of the points if only 1 group is plotted.}

\item{...}{Arguments passed to \code{\link[ggplot2]{aes}}}
}
\value{
A ggplot2 object
}
\description{
Handy scatter plot using ggplot2 that can display regression calculated by lmodel2.
}
\details{

}
\author{
Hao Wu
}
