% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_funs.R
\name{str_to_link}
\alias{str_to_link}
\title{Convert string to HTML link}
\usage{
str_to_link(web.link, display, new.tab = TRUE)
}
\arguments{
\item{web.link}{The embedded web link.}

\item{display}{The text for the hyperlink.}

\item{new.tab}{Should the link be opened in a new tab in the browser?}
}
\description{
Convert string to HTML link
}
