% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sv}
\alias{sv}
\title{A simplified save function using saveRDS().}
\usage{
sv(objname, filename = NULL, path = "./data", compress = T)
}
\arguments{
\item{objname}{expression or char. The object to be saved, use its R object name, e.g., sv(dt).}

\item{filename}{expression or char. The file name of the serialized object. Default to objname, e.g., sv(dt, dt_new_name)}

\item{path}{To which subfolder the data is saved. Default to '../data'. If './', save to current directory.}

\item{compress}{Whether the object should be compressed}
}
\description{
A simplified save function using saveRDS().
}
\examples{
sv(dt) # equals to sv('dt')
sv(dt, dt_new_name, path='../Rdata')
}
