% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version.R
\name{update_version}
\alias{update_version}
\title{Update a Package's Version}
\usage{
update_version(which = NULL)
}
\arguments{
\item{which}{A string specifying which level to increment, one of: "major",
"minor", "patch", "dev". If \code{NULL}, user can choose interactively.}
}
\description{
Our R developers feel the shuffled multiple choice can sometimes be confusing and slows down the process. This function also does not ask to commit files.
}
