% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilr.R
\docType{package}
\name{utilr}
\alias{utilr}
\alias{utilr-package}
\title{utilr: Yu Zhu's data processing routine.}
\description{
The utilr package provides data processing facilities on:
I/O (sv and ld),
NAs filling (fillna),
decimal growth rate (growth),
collaging plots (multiplot),
changing encoding (char2utf8 and char2native)
}
