% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults.R
\name{writeResults}
\alias{writeResults}
\title{writeResults function
print results to uder folder}
\usage{
writeResults(out_dataframe, job_dir, format)
}
\arguments{
\item{results, }{result data}
}
\value{
nothing
}
\description{
writeResults function
print results to uder folder
}
\examples{
writeResults(data = results, format = "tsv", job_dir)
}
