% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyplot.R
\name{dyplot}
\alias{dyplot}
\title{Interactive time series}
\usage{
dyplot(df, vars, dy.label = NA, tz = "UTC", draw.points = FALSE,
  point.size = 2, line.width = 0.5, highlight.individual = FALSE,
  dy.group = NA, ylab = NULL)
}
\arguments{
\item{df}{A dataframe containing variables to plot in a wide format.
Also needs to have a column named \code{date}.}

\item{vars}{The name of variables to plot. If not supplied, all the numeric 
variables will be plotted.}

\item{dy.label}{This specify the column containing additional info to add the 
dygraph. Default is \code{NA}.}

\item{tz}{Time zone of the time series data.#'}

\item{draw.points}{Draw points in the series? The default is \code{FALSE}.}

\item{point.size}{Default value is 2, which is generally appropriate. A value
of 0.5 is very subtle.}

\item{line.width}{Default value is 0.5.}

\item{dy.group}{This specifies the group of the dygraphs. 
If multiple dygraphs are plotted and you want zooming in one of the graphs to be synced with other graphs, 
then set all the dygraphs to the same group.}

\item{ylab}{Name of y axis.}
}
\description{
This function plots an interactive time series using \code{dygraph} package.
}
\examples{

}
\author{
Hao Wu
}
