% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sv}
\alias{sv}
\title{A simplified save function.}
\usage{
sv(x, compress = T, current.dir = F)
}
\arguments{
\item{x}{The object to be saved}

\item{compress}{Whether the object should be compressed}

\item{current.dir}{If True, save to current folder; otherwise, create subdirectory 'Rdata'}
}
\description{
A simplified save function.
}
