% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sv}
\alias{sv}
\title{A simplified save function.}
\usage{
sv(x, path = "../Rdata", compress = T)
}
\arguments{
\item{x}{The object to be saved}

\item{compress}{Whether the object should be compressed}

\item{folder.}{To which subfolder the data is saved. If '.', save to current directory.}
}
\description{
A simplified save function.
}
